+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - add `log` config that allow you to send sms to log directly to use it you should add it to `env`
   as `MSEGAT_LOG = true`.
 - relocate `unicode` and `sender` from `Msegat` class into `MsegatMessage` object.
+- you can specify the time to send the message by calling `timeToExec("YYYY-MM-DD HH:i:ss")` method.